+++ packages/native-core/lib/recipe/text.js
@@ -467,9 +467,22 @@ exports.text = function text(text = "", x, y, options = {}) {
         if (options.underline) {
           const underlineY = y - options.textHeight * 0.1;
           const width = options.lineWidth;
+          ctx.q();
+          if (options.colorModel.xObject) {
+            this._setSeparationColor(
+              options.colorModel.xObject,
+              options.colorModel,
+              true,
+            );
+          }
           ctx
-            .q()
-            .drawPath(x, underlineY, x + width, underlineY, options)
+            .drawPath(
+              x,
+              underlineY,
+              x + width,
+              underlineY,
+              this._devicePathOptions(options),
+            )
             .Q();
         }
       };
@@ -479,9 +492,22 @@ exports.text = function text(text = "", x, y, options = {}) {
         if (options.strikeOut) {
           const strikeOutY = y + options.textHeight * 0.2;
           const width = options.lineWidth;
+          ctx.q();
+          if (options.colorModel.xObject) {
+            this._setSeparationColor(
+              options.colorModel.xObject,
+              options.colorModel,
+              true,
+            );
+          }
           ctx
-            .q()
-            .drawPath(x, strikeOutY, x + width, strikeOutY, options)
+            .drawPath(
+              x,
+              strikeOutY,
+              x + width,
+              strikeOutY,
+              this._devicePathOptions(options),
+            )
             .Q();
         }
       };